backend/requirements.txt
# New FDA Pipeline Requirements

# PDF Processing
PyMuPDF>=1.23.0
pymupdf4llm>=0.0.27

# Text Processing and Tokenization
tiktoken>=0.5.0
langchain-text-splitters>=0.1.0

# HTTP Requests
requests>=2.31.0
httpx>=0.25.0

# Database
SQLAlchemy>=2.0.0
pydantic-settings>=2.0.0

# LLM Integration
langchain>=0.1.0
langchain-core>=1.0.0  # Required for langchain_core imports
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
google-cloud-pubsub>=2.13.0
google-cloud-storage>=2.14.0
google-cloud-aiplatform>=1.38.1
google-api-python-client>=2.110.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
numpy>=1.24.0

# Environment Management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.0.0
pandas>=2.0.0  # For Excel file reading
openpyxl>=3.0.0  # For Excel file writing and reading

# Vector Database
qdrant-client==1.12.1
langchain-community>=0.1.0 

# FastAPI and Web Service
fastapi==0.116.1
python-multipart>=0.0.9
aiofiles==23.2.1
gunicorn==21.2.0
uvicorn>=0.34.0
aiohttp>=3.9.0

# CORS middleware for frontend integration
fastapi-cors==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
PyMySQL>=1.0.0

# SOTA RAG Improvements
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
torch>=2.0.0
scikit-learn>=1.3.0
agno==1.8.0
